# perf: remove localproxy for `frappe.cache`