localproxy is not needed, connection is multitenancy-safe.

### Before

```py
In [2]: %timeit frappe.cache.get_value
1.39 µs ± 10.1 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```

### After (97% better)

```py
In [1]: %timeit frappe.cache.get_value
35.1 ns ± 0.315 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)
```